requirements.txt
streamlit
pandas
numpy
scikit-learn
sentence-transformers
google-generativeai>=0.5.0